Refactoring job screen and list
